samples/email_notification/requirements.txt
Flask
pytz
python-dateutil
requests
